+++ packages/multimethod/src/index.ts
@@ -1,8 +1,9 @@
 import fromMulti from './fromMulti'
+import { Multi } from './internal/types'
 import method from './method'
 import multi from './multi'
 
-export { fromMulti, method, multi }
+export { fromMulti, method, multi, Multi }
 
 export default {
   fromMulti,