+++ packages/multimethod/src/multi.ts
@@ -1,11 +1,11 @@
 import { createMultimethod } from './internal/multimethod'
-import { Multi } from './internal/types'
+import { MultiFn } from './internal/types'
 
 /**
  * Creates multimethod - a function that can dynamically choose proper implementation,
  * based on arbitrary dispatch of its arguments
  */
-const multi: Multi = createMultimethod()()
+const multi: MultiFn = createMultimethod()()
 
 export { multi }
 export default multi