+++ packages/multimethod/CHANGELOG.md
@@ -15,3 +15,9 @@ First stable release.
 ## 1.1.1
 
 - Enabled TS strict checks, fixed TS errors.
+
+## 1.1.2
+
+- Added `Multi` type to exports so it can be extended.
+- Added "defining types" section to README.
+- Added examples for defining types.